+++ apps/word_cloud/setup_viewer.js
@@ -1,4 +1,4 @@
-/* global Coloris, bootstrap */
+/* global Coloris */
 
 import * as constCommon from '../js/constellation_app_common.js'
 import * as constSetup from '../js/constellation_setup_common.js'
@@ -45,7 +45,7 @@ function clearDefinitionInput (full = true) {
             localization: {}
           }
         })
-        previewDefinition(false)
+        constSetup.previewDefinition(false)
       })
   }
 
@@ -86,11 +86,6 @@ function clearDefinitionInput (full = true) {
   })
 
   document.getElementById('promptTextSizeSlider').value = 0
-}
-function createNewDefinition () {
-  // Set up for a new definition
-
-  clearDefinitionInput()
 }
 
 function editDefinition (uuid = '') {
@@ -168,7 +163,7 @@ function editDefinition (uuid = '') {
 
   // Configure the preview frame
   document.getElementById('previewFrame').src = '../word_cloud_viewer.html?standalone=true&definition=' + def.uuid
-  previewDefinition()
+  constSetup.previewDefinition()
 }
 
 function onFontUploadChange () {
@@ -204,28 +199,7 @@ function onFontUploadChange () {
   xhr.send(formData)
 }
 
-function previewDefinition (automatic = false) {
-  // Save the definition to a temporary file and load it into the preview frame.
-  // If automatic == true, we've called this function beceause a definition field
-  // has been updated. Only preview if the 'Refresh on change' checkbox is checked
-
-  if ((automatic === true) && $('#refreshOnChangeCheckbox').prop('checked') === false) {
-    return
-  }
-
-  const def = $('#definitionSaveButton').data('workingDefinition')
-  // Set the uuid to a temp one
-  def.uuid = '__previewWordCloudViewer'
-  constCommon.writeDefinition(def)
-    .then((result) => {
-      if ('success' in result && result.success === true) {
-        // Configure the preview frame
-        document.getElementById('previewFrame').src = '../word_cloud_viewer.html?standalone=true&definition=__previewWordCloudViewer'
-      }
-    })
-}
-
-function saveDefintion () {
+function saveDefinition () {
   // Collect inputted information to save the definition
 
   const definition = $('#definitionSaveButton').data('workingDefinition')
@@ -366,27 +340,20 @@ setTimeout(setUpColorPickers, 100)
 // Add event listeners
 // -------------------------------------------------------------
 
-// Main buttons
-$('#newDefinitionButton').click(createNewDefinition)
-$('#definitionSaveButton').click(saveDefintion)
-$('#previewRefreshButton').click(() => {
-  previewDefinition()
-})
-
 // Settings
 document.getElementById('collectionNameInput').addEventListener('change', (event) => {
   constSetup.updateWorkingDefinition(['behavior', 'collection_name'], event.target.value)
-  previewDefinition(true)
+  constSetup.previewDefinition(true)
 })
 document.getElementById('refreshRateInput').addEventListener('change', (event) => {
   constSetup.updateWorkingDefinition(['behavior', 'refresh_rate'], event.target.value)
-  previewDefinition(true)
+  constSetup.previewDefinition(true)
 })
 
 // Content
 document.getElementById('promptInput').addEventListener('change', (event) => {
   constSetup.updateWorkingDefinition(['content', 'prompt'], event.target.value)
-  previewDefinition(true)
+  constSetup.previewDefinition(true)
 })
 document.getElementById('showExcludedWordsModalButton').addEventListener('click', showExcludedWordsModal)
 document.getElementById('excludedWordsListSaveButton').addEventListener('click', updateExcludedWordsList)
@@ -395,17 +362,17 @@ document.getElementById('excludedWordsListSaveButton').addEventListener('click',
 // Rotation
 document.getElementById('wordRotationSelect').addEventListener('change', (event) => {
   constSetup.updateWorkingDefinition(['appearance', 'rotation'], event.target.value)
-  previewDefinition(true)
+  constSetup.previewDefinition(true)
 })
 // Shape
 document.getElementById('cloudShapeSelect').addEventListener('change', (event) => {
   constSetup.updateWorkingDefinition(['appearance', 'cloud_shape'], event.target.value)
-  previewDefinition(true)
+  constSetup.previewDefinition(true)
 })
 // Text case
 document.getElementById('textCaseSelect').addEventListener('change', (event) => {
   constSetup.updateWorkingDefinition(['appearance', 'text_case'], event.target.value)
-  previewDefinition(true)
+  constSetup.previewDefinition(true)
 })
 
 // Font upload
@@ -415,14 +382,14 @@ document.getElementById('uploadFontInput').addEventListener('change', onFontUplo
 $('.font-select').change(function () {
   const value = $(this).val().trim()
   constSetup.updateWorkingDefinition(['appearance', 'font', $(this).data('property')], value)
-  previewDefinition(true)
+  constSetup.previewDefinition(true)
 })
 
 // Color
 $('.coloris').change(function () {
   const value = $(this).val().trim()
   constSetup.updateWorkingDefinition(['appearance', 'color', $(this).data('property')], value)
-  previewDefinition(true)
+  constSetup.previewDefinition(true)
 })
 
 // Realtime-sliders should adjust as we drag them
@@ -430,7 +397,7 @@ Array.from(document.querySelectorAll('.realtime-slider')).forEach((el) => {
   el.addEventListener('input', (event) => {
     const property = event.target.getAttribute('data-property')
     constSetup.updateWorkingDefinition(['appearance', 'text_size', property], event.target.value)
-    previewDefinition(true)
+    constSetup.previewDefinition(true)
   })
 })
 
@@ -441,7 +408,7 @@ document.getElementById('wordColorMode').addEventListener('change', (event) => {
   } else {
     constSetup.updateWorkingDefinition(['appearance', 'color', 'words'], event.target.value)
   }
-  previewDefinition(true)
+  constSetup.previewDefinition(true)
 })
 
 document.getElementById('colorPicker_words').addEventListener('change', (event) => {
@@ -466,5 +433,6 @@ clearDefinitionInput()
 constSetup.configure({
   app: 'word_cloud_viewer',
   clearDefinition: clearDefinitionInput,
-  loadDefinition: editDefinition
+  loadDefinition: editDefinition,
+  saveDefinition
 })